fix rendering issue when navigating during state change with useCallback
