+++ src/api/GameState.ts
@@ -1,0 +1,6 @@
+export enum GameState {
+  NOT_STARTED = 'NOT_STARTED',
+  START = 'START',
+  ONGOING = 'ONGOING',
+  FINISHED = 'FINISHED',
+}
